src/requirements.txt
Django==1.8.4
South==1.0.2
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.20
celery==3.1.18
cssselect==0.9.1
cssutils==1.0
diff-match-patch==20110725.1
django-celery==3.1.16
django-db-mailer==2.1.0
django-ipware==1.1.0
django-location-field==1.6
django-modeltranslation==0.10
django-pylibmc==0.6.0
django-redis==4.2.0
django-registration==1.0
django-reversion==1.9.3
django-reversion-compare==0.5.5
django-tinymce==2.0.4
httpagentparser==1.7.7
kombu==3.0.26
msgpack-python==0.4.6
premailer==2.9.3
pytz==2015.4
redis==2.10.3
requests==2.7.0
sinchsms==1.0.3
six==1.9.0
stripe==1.25.0
wsgiref==0.1.2
